+++ examples/client-fixed-server.ts
@@ -3,23 +3,20 @@ import * as rpc from '../index'
 
 let port = process.argv[2]
 
-connect(
-    '127.0.0.1',
-    parseInt(port),
-    'secret'
-)
+connect('127.0.0.1', parseInt(port), 'secret')
 function connect(host: string, port: number, token: string) {
     console.log('Token', token)
     class ClientHandler extends rpc.RPCClientHandler {
         onConnect() {
-            this.client.sendMessage({ test: 1 })
-            let obs = this.client.requestObservable('123')
+            const client = this.client as rpc.RPCClient
+            client.sendMessage({ test: 1 })
+            let obs = client.requestObservable('123')
             obs.subscribe(
                 (value) => console.log('Emitted ' + value),
                 undefined,
                 () => {
                     console.log('Observable completed. Closing connection.')
-                    this.client.close()
+                    client.close()
                 }
             )
         }
@@ -28,8 +25,9 @@ function connect(host: string, port: number, token: string) {
             console.log('Client received', data)
         }
 
-        onClose(had_error: boolean) {
+        async onClose(had_error: boolean) {
             console.log(`closed with${had_error ? '' : 'out'} error`)
+            await super.onClose(had_error)
         }
 
         onQuestion() {