+++ examples/fixed-server.ts
@@ -19,10 +19,12 @@ class ClientHandler extends rpc.RPCClientHandler {
 
     onMessage(data: any) {
         console.log('Server received', data)
-        this.client.sendMessage({ test: 'back' })
+        this.client?.sendMessage({ test: 'back' })
     }
-    onClose(had_error: boolean) {
+
+    async onClose(had_error: boolean) {
         console.log(`closed with${had_error ? '' : 'out'} error`)
+        await super.onClose(had_error)
     }
 
     onQuestion() {
@@ -31,8 +33,8 @@ class ClientHandler extends rpc.RPCClientHandler {
 }
 
 const rpcServer = new rpc.RPCServer({ tls: false })
-rpcServer.registerDefaultHandler(
-    (token: string) => (token === 'secret' ? new ClientHandler() : undefined)
+rpcServer.registerDefaultHandler((token: string) =>
+    token === 'secret' ? new ClientHandler() : undefined
 )
 rpcServer.listen(PORT, IP)